requirements-api.txt
# FastAPI and web server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.1.0

# Google OAuth (for Phase 2 implementation)
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# CORS and middleware
python-dotenv>=1.0.0

# Existing dependencies (from requirements.txt)
watchdog==3.0.0
pandas>=2.2.0
pdfplumber==0.9.0
openpyxl==3.1.0
xlrd>=2.0.1
